counter of get_decl
Summary:
How long does a typecheck take? One of our ways of answering is with Counters.ml, e.g.
```
let get_ast_with_error ?(full = false) ctx path =
  Counters.count Counters.Category.Get_ast @@ fun () ->
```

Previously, this `Get_ast` counter covered how long it takes to get the full ASTs for the files we're about to typecheck, plus the non-full ASTs for the files whose declarations we need. But with the recent move over to direct-decl-parser that's no longer the case.

This diff therefore introduces a new `Get_decl` counter to measure how long it takes to fetch all the decls we need.

Reviewed By: shayne-fletcher

Differential Revision: D32587227

fbshipit-source-id: 0164aa70bf4d1da613c7f7b1333ae39447a54292