Use VoidPtr in favor of Int
Summary:
Use `VoidPtr` type instead of `Int` when dealing with pointers.

Adds an `AddOffset` IR that accepts a VoidPtr to do pointer arithmetic (instead of the AddInts we did before).
To support offset-based values for the next diff, we still need to add a separate `AddBase` to treat the VoidPtr as a memory operand for addqrm.

Reviewed By: ricklavoie

Differential Revision: D32569168

fbshipit-source-id: 0ea4a63fbbbc67d84873fd9ca0f254bb0df922fd