Clean up arguments to DepSet.make/singleton/of_list
Summary:
In the previous diff (D31702289), I removed an abstraction, but left some of the arguments to functions which are no longer needed (but were needed to maintain this abstraction) in place.

This diff cleans up those signatures in `typing_deps.ml{i}`. The rest of the diff is updating callsites.

Reviewed By: Wilfred

Differential Revision: D31702289

fbshipit-source-id: e64af7978d9625d4dc35ae5e4ad51b48bd913241