Improve set operations for bespoke arrays
Summary:
Instead of always calling C++ helpers when doing a set operation on
bespoke arrays, try to inline them into the TC. Namely, for struct
dicts (whether concrete or not), we can (usually) do the set inline.

A set operation can be broken down into several steps. First you must
COW the array (if not shared). This can use the existing machinery
we've already invented for irgen-minstr. Then you must look up the
slot in the struct dict that the key corresponds to. This is
accomplished by the new HHIR op StructDictSlot. If the key does not
correspond to a slot, it branches. Now that we have the slot, we must
check if the new value matches the slot's type bound. This is done by
StructDictTypeBoundCheck. It behaves like CheckType, but we might not
know the type bound statically (if we do know, we'll simplify it to
CheckType). We obtain a lval to the old value using
StructDictElemAddr. If it's uninit, there is no previous value, so we
use StructDictAddNewSlot to add the slot to the iteration
order. Otherwise we dec-ref the old value, and use a plain StMem on
the lval to store the new value.

StructDictSet is now no longer needed. The above logic (or punting to
BespokeSet) is all that's needed. These primitives can also be used to
implement BespokeElem (for Dims), as well as IncDec and SetOp in
limited cases (same as in irgen-minstr).

StructDictSlot and StructDictElemAddr can also be used to implement
gets on bespoke struct dicts. This makes LdStructDictElem and
StructDictGetWithColor unnecessary. This simplifies the logic and
makes the business with kDummyException no longer necessary.

Differential Revision: D32528131

fbshipit-source-id: 2a466026ccec20bcfa53ef4b7f0b5d3533318d3d