Allow as expressions to return mutable
Summary:
Allow `as int`, `as bool`, etc to return mutable so that users don't have to write `HH\Readonly\as_mut` for simple cases. Any integer is "mutable" by default.

Turn it on in HHVM first, and then we can turn it on in the typechecker once HHVM is released after.

Perflabs after last two diffs:
https://our.intern.facebook.com/intern/servicelab/1001705496/
https://our.intern.facebook.com/intern/servicelab/701703785/
https://our.intern.facebook.com/intern/servicelab/601705865/
https://our.intern.facebook.com/intern/servicelab/901706131/
https://our.intern.facebook.com/intern/servicelab/901706130/
https://our.intern.facebook.com/intern/servicelab/402277113/
https://our.intern.facebook.com/intern/servicelab/402277112/
https://our.intern.facebook.com/intern/servicelab/502277367/

Reviewed By: oulgen

Differential Revision: D32573604

fbshipit-source-id: 5635fbae069ea4492c6d932ae1dc3dec35919707