Add optimization pass to lower bespoke array accesses
Summary:
During irgen time we'll emit some specialized sequences for some
struct dict bespoke accesses. However, some arrays might only become
apparent as struct dicts after optimization passes. The most common
reason is if frame-state drops state due to loops. Load-elim is needed
to recover this.

Since it's desirable to use these specialized sequences as much as
possible, add an optimization pass to lower a generic bespoke array
access into a specialized one (if applicable). The logic is
essentially the same as the irgen logic, except we need to modify the
cfg manually. The pass is analagous to a simplifier rule, except we
cannot actually use the simplifier because we need to add control
flow.

Right now only BespokeGet and BespokeGetThrow are supported. The set
operations could be dealt with as well, but they're more
tricky. During irgen, we might emit side-exits which use the generic
ops, and we want to leave those untouched. We'd need to add some logic
to determine if something is a side-exit or not.

Why do this during irgen at all, instead of just letting the optimizer
deal with it? The code generated by the optimization pass isn't quite
as good as the irgen output. This is because irgen has more context,
while the optimizer must maintain the semantics of the original op. We
hope later optimization passes do their best.

Differential Revision: D32592019

fbshipit-source-id: 38927a45067b8b81f78a785ec8c1b12c134547a4