Allocate evictable values in the evictable allocator
Summary:
We now have two allocators: one for evictable values and one for non-evictable values. We also have a flag that let's us know whether the value we are inserting is evictable or not.

If we we are inserting an evictable value, we should mark it as such in its header, and allocate the bytes in the allocator for evictable values.

Differential Revision: D31109903

fbshipit-source-id: 1325ca4c6fef99a1b88c17ec8c03cf4b6c31ede5