Remove in case
Summary:
`// FALLTHROUGH` is interpreted as a comment outside switch statements and the `in_case` field of the environment is only used to decide on how to handle `Fallthrough` statements. In effect, `in_case` is always `true` when a `Fallthrough` statement is encountered during typechecking.

This diff removes the `in_case` field and manipulates the environment while typechecking `Fallthrough` as if `in_case` was set to `true`.

Reviewed By: vsiles

Differential Revision: D32401265

fbshipit-source-id: d1580fcb2fa6b8cd66934aeecfb3a97b7d7c82f2