remove redundant refinement code
Summary: There was `get_instance_var` which returns an id or raises an exception, and a companion `is_instance_var` to check whether `get_instance_var` will succeed. There was also `make_a_local_of` that returns an `id option`. Since the two were almost identical, I removed the is/get functions and use the make function in their place. The one difference is that `get_instance_var` considered `$this` to be an instance var and `make_a_local_of` didn't. Although the reason for the difference is unclear, we use a flag to exactly replicate the behaviour.

Reviewed By: CatherineGasnier

Differential Revision: D32614745

fbshipit-source-id: cf44c87a49623564bb86f73f08aa7cd992ed78d5