Fix bespoke lowering when we know element is present
Summary:
BespokeGet's return type might not include TUninit if we know the
element will always be present. In this situation, when lowering a
BespokeGet to a specialized sequence, we should not include the
"Uninit" branch of the sequence. If we do so, we'll union TUninit into
the result, incorrectly widening the type. Only emit the branch if the
lookup can fail. If it can't, emit an Unreachable instead. This will
be optimized away later.

Reviewed By: oulgen

Differential Revision: D32692241

fbshipit-source-id: 7ff772f933ed2ffd7488550f53b5783497a467a9