Differentiate /A from ^ for regex
Summary:
In PCRE, `/^def/` differs from `/def/A` in that `/A` is still able to match in the middle of a string if an offset if provided. They currently are both treated like `^`.

per pcreapi [docs](https://www.pcre.org/original/doc/html/pcreapi.html):
```
  PCRE_ANCHORED
If this bit is set, the pattern is forced to be "anchored", that is, it is constrained to match only at the first matching point
in the string that is being searched (the "subject string").

...
The "start of line" metacharacter (^) matches only at the start of the string
```

Reviewed By: fredemmott

Differential Revision: D32691175

fbshipit-source-id: 65dbdac38a7e729b7574ef15487d9392657cbf95