Improve the error messages on dict/darray value incompatibility
Summary:
We have an error that reads:

```
The values of this `Map` are incompatible (Typing[4110])
```

This is likely due to the previous usage of Hack Collections in the codebase. These days, we no longer use `Map` as frequently. Instead of using the term `Map` in the error, use the name of the collection in question.

Thus, if the code in question is a darray, use darray in the error. Same for dict.

Reviewed By: Wilfred

Differential Revision: D32609068

fbshipit-source-id: 78148de8ae5c11a152273d189fa596142c494db5