Never consider Optimize translations as incomplete
Summary:
If there are bindjump or fallback instructions in the main code area, the unit
is considered incomplete and not a candidate for inlining.  This diff disables
this logic for Optimize regions: we should trust the profile data that was used
to form them, and any side exit in the region should be rare, otherwise such
path would have been included in the region.

Reviewed By: jano

Differential Revision: D32295598

fbshipit-source-id: 22a337a906f91e28845f0e3d90158ad271f397a3