Fix the fix for bespoke lowering
Summary:
I forgot that slotAlwaysPresent only checks if the slot is
required, not that if a slot exists. IE, if you just pass TInt (any
slot), and the layout has all required values, it will return true. Of
course, this doesn't mean the slot lookup will succeed, because they
key might not be present at all. Instead of trying to infer the
presence of the key, just follow the original type of the dst.

Differential Revision: D32714139

fbshipit-source-id: d0954319684d3117b921f5bc7d53ef2dce6fdf92