Fix property profile serialization
Summary:
One isn't allowed to iterate over a concurrent_hash_map
concurrently with insertions (or deletions) to the same map. There's
no real guarantees if you do so. I don't think you can get a crash
with the current implementation, but you can see the same entry twice,
or skip over entries.

Instead of using iteration over the map, keep a separate vector of the
keys. The vector is protected by a normal lock. We only have to insert
into it the first time we see a key (so contention should quickly
become rare). When serializing, make a copy of the vector, and iterate
over that. Since the copy is local, we don't have to worry about
concurrent modifications. For each key in the vector, do a normal get
on the map (safely concurrent with other accesses).

Add some checking for duplicates before serializing, and improve
assertion on the deserializing side.

Reviewed By: ottoni

Differential Revision: D32714650

fbshipit-source-id: e86eabe7fd879bf629fe512463ddb9e643889af2