partially respect cancellation requests
Summary:
It's hard to properly respect cancellation, i.e. interrupt a request already underway. Our typechecker and RPC APIs simply aren't set up for that.

This diff does the next best thing. It avoids us embarking upon handling of a request if that request has already got a cancellation request. It's in a new method `ClientLsp.cancel_if_has_pending_cancel_request`

This can be a racey thing! For instance, a cancellation request might be on the stdin of our jsonrpc-reader but not yet ingested at the time we embark upon handling the request, and hence we won't discover in time. I think that's just the inherent nature of cancellation.

Reviewed By: CatherineGasnier

Differential Revision: D32670024

fbshipit-source-id: 5e03ac0fe459061c534b5cfe50d807ca80201dc6