Improve refineTmps
Summary:
The refineTmps pass attempts to rewrite the source of a
CheckType/AssertType with it's dest whenever possible. It can only do
so if the dest is available (and there hasn't been a Call
instruction). That is, it's best effort. Afterwards there may be uses
of the source after the CheckType/AssertType.

reorderCheckTypes was relying on a stronger invariant. That is, the
pass is only correct if refineTmps is completely successful in
rewriting all downstream uses of the source.

Improve refineTmps to make this invariant guaranteed. The logic is
similar to the original version, but if the dest is not available,
insert a phi between the dest and the old source. The rewrite then
proceeds with this new SSATmp. While we're at it, instead of special
casing CheckType/AssertType, use all pass through instructions.

Not only does this make reorderCheckTypes correct, it can improve
code-gen. It more reliably ensures that only one logical copy of a
value is live at a time.

Reviewed By: ottoni

Differential Revision: D32692316

fbshipit-source-id: 258e908c96fb643fba707114bc3874d9d9a07b1f