Fix reachability analysis in simplify
Summary:
The simplify pass tracks which blocks are reachable (because we do not
want to run the simplifier on unreachable blocks). However, its
analysis is too simplistic. It visits each block in RPO order. If the
block is reachable, mark its successors as reachable.

However, we might visit a block's predecessor *after* we've processed
the block. This can happen during CFG optimizations. Instead of just
using RPO order, use a worklist. Only schedule a block once we process
a predecessor

Reviewed By: oulgen

Differential Revision: D32692298

fbshipit-source-id: ead12db16dfb85fb3fc18a26ee93001942ed0bc3