Improve error messaging when passing a readonly value to an inout arg
Summary:
When you pass a readonly (inferred) value to a inout argument, HHVM will throw a nonsensical emitter error due to us creating a ReadonlyExpr(Inout($arg)) instead of an Inout(ReadonlyExpr($arg)). This gets past our inout expression check because readonly_check happens after rust parsing errors.

Hack still throws the correct typechecker error (along with a weird NastCheck error) but if you happen to be doing something untyped/evaling it, you'll run into this bug, so it's better to error right away and consistently.

We fix this by erroring directly in readonly analysis when trying to use inout on a readonly value in a call.

Reviewed By: KendallHopkins

Differential Revision: D32773556

fbshipit-source-id: 5aef345fdb87b4426de7989cdadafbffd381ab05