supportdynamic as a generic type alias
Summary:
An alternative to a single type `supportdynamic` (the supertype of all non-null types that support dynamic) is an opaque generic type `supportdyn<t>` that semantically is equivalent to `(t & dynamic)`. This lets us describe structured types such as open shapes and function types which by default do not support dynamic, simply by intersecting them with `dynamic`. It's enough to define
```
newtype supportdyn<+T> as T = (T & dynamic);
```

Declarative subtyping rules for `supportdyn<_>` as as follows.
```
-------------------------
supportdyn<t> <:D dynamic

------------------
supportdyn<t> <: t

t <: u    t <:D dynamic
-----------------------------
t <: supportdyn<u>
```
This is just the interpretation of `supportdyn<t>` as the intersection type, but with the right-hand side of the intersection only accessible through dynamic-aware subtyping.

Algorithmic implementation of these rules is a bit messy. The rule `supportdyn<t> <: t` comes for free from the bound on the `newtype`.

We remove `Tsupportdynamic` and `Hsupportdynamic` but instead define the original `supportdynamic` type as an alias:
```
type supportdynamic = supportdyn<nonnull>;
```

One other change to subtyping is that we deal specially with `t <:D dynamic` rather than `t <: supportdynamic`. This in turn affects constraints in the environment. Strictly speaking, we should have separate lower and upper bounds for type variables wrt both `<:` and `<:D`. That's a bigger change. For now, when adding a constraint `<:D dynamic` we instead add `<: ?supportdynamic`, which is equivalent.

Future diff will express SDT function types simply as `supportdyn<(function(int):bool)>`, for example.

Reviewed By: francesco-zappa-nardelli

Differential Revision: D32429192

fbshipit-source-id: 32311ed48c55a148edf2c82cca6e4dcd2911c0c1