Cleanup: remove unsafe Str::as_ref() returning &str
Summary:
Frankly I'm surprised that rustc allows this -  `Str` is just a type alias for `Slice<u8>` so you end up with the same named method defined with a different return type.

Also rename `Str::as_str()` to `unsafe_as_str()` - it's a fundamentally unsafe operation so let's be clear about it and make it easier to find (since there are a lot of valid `as_str()` methods).

Reviewed By: shayne-fletcher

Differential Revision: D32698285

fbshipit-source-id: 4d18d5a28e1a57715d4cc3590c936464eb9bd843