Check for object property writes when write_props capability is missing, not just when function is pure
Summary:
Currently the coeffects check only looks for object property writes inside pure functions; however, this is not as restrictive as it should be - any function lacking write_props should be checked. This diff:

1. makes that change
2. updates tests to test the reverse: that no errors are emitted when write_props *is* available
3. transitions from checking whether a function is pure to whether it contains a particular capability and populates this pattern with write_props (in subsequent diffs, logic for read globals and access globals is populated)

NOTE: To land this safely, we need to check if we are in the typechecker (as opposed to hhvm runtime). If someone commits a new WWW error not amended in D32268019 when this is landed in hhvmtrunk but before the hack release, they will not catch it in development and could land something that breaks hhvm trunk.

This diff also preserves all the logic around is_pure to allow us to run the check in the runtime when the function is pure (as before). This way we can keep the existing hhvm tests and not reverse the existing behavior. When we release the changes to the runtime (after the hack release with this stack of diffs is completed) then the is_pure logic can be cleaned up

Reviewed By: oulgen

Differential Revision: D31979525

fbshipit-source-id: da5cccdf68c143c0eb8cc9ad8c5422b282f676dc