Fix HSL locale tests on MacOS (#167)
Summary:
This test was previously testing undefined behavior.
- specify the encoding to make it defined where practical
- delete tests of other stuff like how sort order handles characters
  that are not valid in the current encoding.

The entire format of locale strings  is implementation-defined in POSIX;
`foo_BAR[.ENCODING]` is just a common convention, and there does not
appear to be a convention for default encodings.

fixes https://github.com/hhvm/hsl/issues/165

I filed that issue as I was worried it might be handled inconsistently;
it's not, we're correctly detecting it as a multi-byte locale here:

https://github.com/facebook/hhvm/blob/7e621926354146812f1d9532614947270e1e7ff2/hphp/runtime/base/locale.cpp#L130-L139

I think this is the better thing to do as:
- hack stuff will meet user expectations when encoding is specified
- it would be pretty reasonable to lint to require an explicit encoding
- if C library calls and subprocesses interpret the locale differently to how
  Hack/the HSL do, that seems like a potential source of bugs

Pull Request resolved: https://github.com/hhvm/hsl/pull/167

Reviewed By: oulgen

Differential Revision: D32804489

fbshipit-source-id: 5442eb11dbbaf61653e4ae2ff1ee38f0f69e3fc5