Replace verify_param_type_ts with verify_param_type based on class decls
Summary:
When we compile a function:

```
function foo(A $a): void {}
```

We will produce bytecode that checks the type of the arguments that we get passed in to ensure that the types are aligned. For instance, these bytecodes would check to determine that the following:

```
foo(new NotA());
```

Would be an invalid call that would emit a runtime type hint violation exception. These bytecode type checks become more interesting when the type that we are checking has generics.

```
final class A<T> {}
function foo(A<int> $a): void {}
```

Typically, generics are erased at the runtime. Thus, the following would occur:

```
foo(new NotA()); // This results in a runtime type hint violation exception because we are not passing an object of type A
foo(new A<string>()); // This is a typechecker but crucially, not a runtime typehint violation
foo(new A<int>()); // This is accepted by both the typechecker and the runtime
```

As a way to bridge the gap between the typechecker and the runtime, reified generics are a way of more deeply enforcing the generics in a class.

```
final class A<reify T> {}
function foo(A<int> $a): void {}
```

Now, those calls have the following behavior:

```
foo(new NotA()); // This results in a runtime type hint violation exception because we are not passing an object of type A
foo(new A<string>()); // This is a typechecker and now also a runtime exception!
foo(new A<int>()); // This is accepted by both the typechecker and the runtime
```

In order to validate the generics, we emit different bytecode when compiling the function. We emit `VerifyParamTypeTS` bytecodes that also does a type structure comparison. This check is naturally more expensive than the check we would emit in the case where we don't need to check the generics `VerifyParamType`.

However, due to the file by file emission/compilation strategy we currently pursue, when we compile a file, we do not know if class `A` contains reified generics or not. If class `A` just contained erased generics, we could get away with only emitting `VerifyParamType` bytecodes. But at compilation time, we do not have that information, so we currently always emit the more expensive `VerifyParamTypeTS` bytecodes. While this is potentially optimizable, our current strategy is conservative,

https://www.internalfb.com/code/fbsource/[2874ec6888c5f267a8535d3f2333203b704726e5]/fbcode/hphp/hhbbc/interp.cpp?lines=4838-4843

If we could determine, at compile time, whether a class contained any reified generics, we could potentially emit more efficient bytecode upfront, instead of relying on HHVM to potentially optimize this. Thus, our strategy is the following:

1. When we compile a function/method, scan it's arguments
2. See if any of the arguments are typed and those types have generics (`A<int>`)
3. Do a call to the decl_provider to determine whether the type contains any reified generics (`decl_get(A)`)
4. If the class has no reified generics, emit the cheaper `VerifyParamType`
5. If we do not get the decl back, or there are reified generics on the class, emit the more expensive `VerifyParamTypeTS` as we do now

The benefit is that if the decl_provider were to fail or if we cannot find the class for some reason, the code could still work. Thus, this is an optimization that helps in the case where decls are found, but does not cause the world to explode if decls were not found.

Reviewed By: shayne-fletcher

Differential Revision: D32129779

fbshipit-source-id: a14f3e30a6e5404ca032eb86619d043f63ed1f5e