Add constructor and "copy constructor" for Context object to persist state of bit field
Summary:
When using bit fields instead of boolean fields in the `Context` object to represent availability of capabilities, we can't use the `..` notation to initialize a new `Context` with the rest of the  fields from another instance. But for each next recursive call in `visit_fun_` or `visit_method` or `visit_fun_def` we need to preserve some previous values (e.g. `is_typechecker`) while overwriting others (e.g. `has_defaults`).

Setting the changes on the existing `Context` object is incorrect - it is the cause of the bug here: https://www.internalfb.com/diff/D32012420?transaction_fbid=264866185670197. Additionally making this object not mutable would help eliminate this class of bugs, but the visitor framework won't allow it.

So at least having a constructor and a sort of "copy constructor" to create a new `Context` is helpful. I wrote `from_context()` instead of using `#[derive(Copy, Clone)]` to achieve this because the `bitflags!` macro obscures some of the implementation of the `ContextFlags` struct - it looks like these traits are already implemented elsewhere.

Reviewed By: oulgen

Differential Revision: D32655770

fbshipit-source-id: 172879815615098fca1bf50ab469634eb24535d7