Update error pos
Summary:
User reported that some errors reported were hard to diagnose. Part of this may be due to the way that VSCode does underline.

In the example given, the user had a bad splice which caused Hack to report an error on the desugared dictionary of `The values of this Map are incompatible`.

However, due to the way that we associate positions with the desugared Expr nodes, we report the errors on the desugared code as the entirety of the Expression Tree. Thus, when users are hovering over the red swiggle in VSCode, they see errors on the invisible desugared code which obscures the real underlying issue.

In this diff, we change the desugaring and virtualization process such that if we fabricate a node that doesn't have an exact position that it is fabricated from, it uses the position of the hint instead of the position of the whole expression. This way, users may get better feedback when they hover over more specific red underlines.

Reviewed By: Wilfred

Differential Revision: D32608200

fbshipit-source-id: 3be5f000f5ecc36cc1f6ad1f051ed315f8c9ce33