Fix unions and intersections of specializations and constant data in Type
Summary:
If you union together two Types, one of which has a constant value,
we'll drop the constant value if there's not an exact match. This can
cause non-monotonicity. If we have two types, both of the same
specialization, and union them together, we'll obtain the same
specialization. However, if we refine of them to a constant value, the
union will now not have any specialization, which is a larger type!

When unioning or intersecting Types, if one is a subtype of the other,
we can just return the left or right hand side, without further
operations. Thus refining a type to a constant will still produce the
same union, as the constant is a subtype of the specialization.

Differential Revision: D32867804

fbshipit-source-id: 30814145c8128745037455d12047c9cc9e522a4c