Refactor out some logic to reuse elsewhere
Summary:
In the previous diff, we did the most basic and conservative use of decls in compilation for when to replace `VerifyParamTypeTS` with `VerifyParamType`.

There are more places where we can reuse the decl checking logic used. This diff factors out that logic so it can be more easily called into more other places in the code base.

Differential Revision: D32224679

fbshipit-source-id: 4ca98e49439d9cc0a7b4b1036b25db011cf1e562