Check decl even if generic used is reified
Summary:
In the previous diff, we began emitting the `VerifyParamType` bytecode instead of the more expensive `VerifyParamTypeTS` bytecode when the class we were checking had no reified generics.

In this diff, we expand the scenarios where we check the class to determine whether it has reified generics.

Previously, we would not check the class `A` in the following situation:

```
function test<reify T>(A<T> $_): void {}
```

We see that the type parameter in `A<T>` is reified in the function declaration: `test<reify T>`, thus, we would not check to determine whether `A` had reified generics. Instead, we would always emit `VerifyParamTypeTS`.

However, it should be correct, if `A` did not have reified generics, to simply emit `VerifyParamType`. For instance, if the definition of the `A` class were:

```
class A<T> {}
```

Thus, this diff adds an additional check. If any parameter has a generic parameter, regardless of whether any of the generics are reified generics in the current context, we will check the type to determine whether it is a reified class. If it isn't, we will emit the cheaper `VerifyParamType` bytecode instead of the `VerifyParamTypeTS` one.

Differential Revision: D32260580

fbshipit-source-id: d498f9212d794c9cefd390e2a86702177ba50b9b