SimpleParser: Change to nonrecursive parsing algorithm
Summary:
Change SimpleParse parsing algorithm to a nonrecursive version that maintains state on a separately maintained stack.

This was originally motivated to eliminate performance difference between gcc/clang where clang inline recursive functions a lot less aggressively and the new version of the code is faster (and smaller) on both compilers.

Reviewed By: ottoni

Differential Revision: D31517969

fbshipit-source-id: 71daffe220c985056ba3003e42bad2aeed329d46