SimpleParser: Reject any number starting with 0 not just in arrays
Summary:
JSON/ECMA-404 do not accept leading zeros for any number, regardless of whether they are in an array or not. Simplify the code by removing that special case where they were accepted outside of arrays.

While on it also fix a bug where a number with a minus sign and leading zero was incorrectly accepted.

(I am doing this to simplify the code in the other changes in this stack)

Reviewed By: jano

Differential Revision: D31517970

fbshipit-source-id: be665f5055949ad1880c642cd21a92d1708e763c