Add crc generation to hh_compile
Summary:
Added a new mode to hh_compile so that instead of printing the bytecode output for files it prints a CRC of that bytecode.  This is useful to be able to quickly compare if bytecode output has changed for a large number of test files.

Note that because this produces a CRC based on the textual output of the bytecode printer it may miss binary-only differences. This may be worth adding in the future.

Reviewed By: shayne-fletcher

Differential Revision: D32815446

fbshipit-source-id: 47443a11f6a97a1e1fda7b98db46f02308da0de2