remove varray and darray
Summary: Remove darray, varray, and varray_or_darray constructors from hints and types. `Hvarray` and `Hdarray` were only constructed by the parser as part of the type hint on an `is` or `as` expression, which were then always type errors. At all other places, the parser/lowerer turned "varray" and "darray" inside of types into "vec" and "dict". This makes all type references to "varray" and "darray" become "vec" and "dict" (which does lose the type error for `is varray<_>`, but that seems harmless, since `is vec<_>` is allowed. The only place `vector_or_darray` was created is where the special type "PHPism_FIXME_Array" was detected, and so we change it to `vec_or_dict`.

Reviewed By: andrewjkennedy

Differential Revision: D32895824

fbshipit-source-id: a5f9529011bd4d41c8a676de9bd8d7300acb4901