Correctly check SDT lambda under dynamic assumptions
Summary: Checking of SDT lambdas under dynamic assumptions was broken, because the environment used to kick off the check was captured *after* the first check under static types, which leaves continuation typings as they are at the end of checking the body! Instead, we should capture the environment as set up just *before* the first check.

Reviewed By: vassilmladenov

Differential Revision: D32917321

fbshipit-source-id: 8ee8e4c508be7efebf28297ac0c9123362bb7bfc