When determining which coeffects checks to run, check if the capability is contained by context, is a subtype of a capability, or is an exact match
Summary:
This diff is a small design improvement.

To determine which checks to run (write_props vs. read_globals vs. globals, etc) we used mappings from capability -> contexts it is contained in (in the form of helper functions).

Searching through the hints in a function definition and calling the appropriate helper for each capability led to less code, but means that if we have a new context in the future, we have to add it to every capability's list of contexts.

Instead, we can use a mapping from context -> capabilities it contains (again in the form of helper functions).

This diff still searches through each hint in the function definition but now simply calls the appropriate helper for each context. When a new context is added, we just add its mapping to capabilities (and helper function call) - all in one centralized place.

Reviewed By: oulgen

Differential Revision: D32193347

fbshipit-source-id: a5a9061594214f0200175f3433c1f62945e0e264