load from naming sqlite in sandcastle as well
Summary: In sandcastle we download the saved state seperately in a sandcastle step and pass in a json info blob that tells us where all the artifacts are. This diff updates that process to read from the naming sqlite table in the hack/64 saved state if it exists. We need this because we need a sqlite table with decl hashes to fetch remote decls.

Differential Revision: D32956368

fbshipit-source-id: ce65d17f92f89d68b3b52dfed6e0038e6159a841