Replace `\0` with `|` in mangleUnitSha1()
Summary:
This is part of a stack of diffs breaking `mangleUnitSha1()` apart. In native Facts, I use `mangleUnitSha1()` to mutate the cache keys for source files, and to determine where to store the Facts DB. I want to be able to look inside and see what options are causing us to hit a new cache key or store the DB in a new place.

I want to log the inputs to `mangleUnitSha1()`, but our logging APIs get screwed up by the null terminators here. `HPHP::Logger` and `TRACE()` both use C strings.

I don't think there's a reason to prefer a null terminator, or any unprintable character, here.

`mangleUnitSha1()` is guaranteed to be different in every release of HHVM, so there are no stability concerns with changing its format.

Reviewed By: ricklavoie

Differential Revision: D32976393

fbshipit-source-id: 7bb41ab62dd72464e95004c77619569dc69dba1c