Propagate readonly attribute for properties on ancestors
Summary:
When doing a property lookup, if we don't find it on the base class, we check the ctx class if its an ancestor to see if it has a private property of the same name. If it does, we return that.

We also need to return the readonly-ness of that property if we find it.

Reviewed By: ricklavoie

Differential Revision: D33010322

fbshipit-source-id: 060b10909bb15ecf07f5447f01a6c3da382a76fd