Stop building a hash table for APC dicts
Summary:
The only operation that APCArray supports is serialization and deserialization: we store (key, value) pairs in it at construction and read them out, in order, at fetch. There's no reason to use a hash table here. We don't need to de-dupe keys, because we construct these arrays from Hack dicts, which have unique keys.

For the new layout, we use a simple array of values for vecs and keysets; or, an array that alternates (key, value) for dicts.

Reviewed By: ricklavoie

Differential Revision: D32706270

fbshipit-source-id: 45585760f4ac242f0ae337cd72d1c2f4d0d99518